+++ src/main/java/com/wandted/matitnyam/restaurant/dto/RestaurantListResponse.java
@@ -3,7 +3,7 @@ package com.wandted.matitnyam.restaurant.dto;
 public record RestaurantListResponse(
         Long id,
         String name,
-        Long longitude,
-        Long latitude
+        Double longitude,
+        Double latitude
 ) {
 }
